Extracted MissingXmlDocHighlightingTestsBase class, added new DefaultSettingsTests, fixed non side-effect free changing of settings (in WithProject) with ExecuteWithinSettingsTransaction(), moved .gold files from sample project to the test project (next to the test fixtures)
