+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/ClassesAndMembersWithDocs.cs.gold
