+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/ClassesAndMembersWithoutDocs.cs.gold
