+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/DerivedClassWithoutDocs.cs.gold
