+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/InterfacesWithoutDocs.cs.gold
