+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/PartialClassWithDocsInOneFile_A.cs.gold
