+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/PartialClassWithDocsInOneFile_B.cs.gold
