+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/PartialClassWithoutDocs_A.cs.gold
