+++ XmlDocInspections.Plugin.Tests/IntegrationTests/BasicTests/PartialClassWithoutDocs_B.cs.gold
