+++ XmlDocInspections.Plugin.Tests/IntegrationTests/DefaultSettingsTests.cs
@@ -1,0 +1,18 @@
+﻿using JetBrains.Application.Settings;
+using NUnit.Framework;
+
+namespace XmlDocInspections.Plugin.Tests.IntegrationTests
+{
+    public class DefaultSettingsTests : MissingXmlDocHighlightingTestsBase
+    {
+        protected override void MutateSettings(IContextBoundSettingsStore settingsStore)
+        {
+        }
+
+        [Test]
+        public void TestClassesAndMembersWithoutDocs()
+        {
+            DoNamedTest2();
+        }
+    }
+}
